+++ src/main/java/pokeraidbot/commands/ConfigAwareCommand.java
@@ -17,11 +17,7 @@ import java.util.concurrent.*;
 public abstract class ConfigAwareCommand extends Command {
     protected final ServerConfigRepository serverConfigRepository;
     protected final CommandListener commandListener;
-    private final LocaleService localeService;
-    // todo: pass thread pool as an argument, so we have the same pool for all worker threads
-    protected static final ExecutorService executorService = new ThreadPoolExecutor(100, Integer.MAX_VALUE,
-            120L, TimeUnit.SECONDS,
-            new SynchronousQueue<>());
+    protected final LocaleService localeService;
 
     public ConfigAwareCommand(ServerConfigRepository serverConfigRepository, CommandListener commandListener,
                               LocaleService localeService) {