+++ CHANGELOG.md
@@ -13,6 +13,13 @@ should be put in a read-only channel
 * !raid mapinchat - raid map forced in server chat, despite server settings (meant for admins in servers with such config)
 * Minor text changes here and there
 * Possible to change time for a raid group: !raid change group (time) (gym)
+* !raid list should show next ETA
+* Clean up signups for expired group
+* Persistent tracking
+* User configuration, stored in database. 3 available trackings, and locale.
+* User can set their own locale
+* Move time for raid group: !raid change group (time) (gym)
+* Move attaching to overview to its own listener. Doesn't belong in EventLoggingListener.
 
 0.9.1 (2017-10-14)
 ======